Keep project path helpers in their own file
We have a bunch of these path helpers so splitting them into their own
files seems like a good plan.
