+++ spec/helpers/application_helper_spec.rb
@@ -92,44 +92,6 @@ RSpec.describe ApplicationHelper, type: :helper do
     end
   end
 
-  describe "project paths" do
-    before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
-
-    context "when the project is a voluntary conversion" do
-      it "returns the correct paths" do
-        project = create(:voluntary_conversion_project)
-
-        expect(helper.path_to_project(project)).to eq conversions_voluntary_project_path(project)
-        expect(helper.path_to_project_task_list(project)).to eq conversions_voluntary_project_task_list_path(project)
-        expect(helper.path_to_project_information(project)).to eq conversions_voluntary_project_information_path(project)
-        expect(helper.path_to_project_notes(project)).to eq conversions_voluntary_project_notes_path(project)
-        expect(helper.path_to_project_contacts(project)).to eq conversions_voluntary_project_contacts_path(project)
-
-        expect(helper.path_to_team_lead_project_assignment(project)).to eq conversions_voluntary_project_assign_team_lead_path(project)
-        expect(helper.path_to_regional_delivery_officer_project_assignment(project)).to eq conversions_voluntary_project_assign_regional_delivery_officer_path(project)
-        expect(helper.path_to_caseworker_project_assignment(project)).to eq conversions_voluntary_project_assign_caseworker_path(project)
-        expect(helper.path_to_assigned_to_project_assignment(project)).to eq conversions_voluntary_project_assign_assigned_to_path(project)
-      end
-    end
-
-    context "when the project is a involuntary conversion" do
-      it "returns the correct path" do
-        project = create(:involuntary_conversion_project)
-
-        expect(helper.path_to_project(project)).to eq conversions_involuntary_project_path(project)
-        expect(helper.path_to_project_task_list(project)).to eq conversions_involuntary_project_task_list_path(project)
-        expect(helper.path_to_project_information(project)).to eq conversions_involuntary_project_information_path(project)
-        expect(helper.path_to_project_notes(project)).to eq conversions_involuntary_project_notes_path(project)
-        expect(helper.path_to_project_contacts(project)).to eq conversions_involuntary_project_contacts_path(project)
-
-        expect(helper.path_to_team_lead_project_assignment(project)).to eq conversions_involuntary_project_assign_team_lead_path(project)
-        expect(helper.path_to_regional_delivery_officer_project_assignment(project)).to eq conversions_involuntary_project_assign_regional_delivery_officer_path(project)
-        expect(helper.path_to_caseworker_project_assignment(project)).to eq conversions_involuntary_project_assign_caseworker_path(project)
-        expect(helper.path_to_assigned_to_project_assignment(project)).to eq conversions_involuntary_project_assign_assigned_to_path(project)
-      end
-    end
-  end
-
   describe "#enable_google_tag_manager?" do
     context "when not in production" do
       it "returns false" do
